fix: rename DownloadManegement.tsx (typo) and fix models.ts JSDoc params
- DownloadManegement.tsx -> DownloadManagement.tsx
- models.ts extractModelName/extractModelRepo documented modelId;
  the actual parameter is model
